+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -68,22 +68,7 @@ public class ImageController {
         return "images/image";
     }
 
-    //Adding comments
-    @RequestMapping(value="/image/{id}/{title}/comments", method = RequestMethod.POST)
-    public String addComments(@PathVariable("id") Integer id, @PathVariable("title") String title, Comment newComment, HttpSession session){
-        User user = (User) session.getAttribute("loggeduser");
-        newComment.setUser(user);
-        Image image = imageService.getImageById(id);
-        newComment.setImage(image);
-        System.out.println("CREAM:"+newComment.getUser().getUsername());
-        System.out.println("CREAM:"+newComment.getText());
-        System.out.println("CREAM:"+newComment.getImage().getTitle());
-        Date date = new Date();
-        newComment.setCreatedDate(date);
-        System.out.println("CREAM:"+newComment.getCreatedDate().toString());
-        commentService.createComment(newComment);
-        return "redirect:/images/"+id+"/"+title;
-    }
+    //Adding comments function moved from here to CommentController
 
 
     //This controller method is called when the request pattern is of type 'images/upload'
